Use 3-letter month abbreviations in news dates for alignment
Changed formatDate to use 'short' month format (Jan, Feb, etc.) instead of
full month names to ensure consistent date alignment in the news section.

+++ sanity/lib/formatDate.ts
@@ -1,6 +1,6 @@
 const formatDate = (inputDate: string) => {
     const date = new Date(inputDate);
-    return date.toLocaleString('default', { month: 'long', year: 'numeric' });
+    return date.toLocaleString('default', { month: 'short', year: 'numeric' });
 }
 export const formatDateLong = (inputDate: string) => {
     const date = new Date(inputDate);
